backend/requirements.txt
# AI Buddy — Python Backend Dependencies
# Install: pip install -r requirements.txt

# Web server & WebSocket
fastapi==0.115.0
uvicorn[standard]==0.30.0
websockets==12.0

# Speech-to-Text (local, fast)
faster-whisper==1.0.3

# LLM — Google Gemini (default)
google-generativeai==0.4.1
openai==1.14.0
psutil==5.9.8
duckduckgo-search==5.3.0

# TTS — Microsoft Edge TTS (free, natural voices)
edge-tts==6.1.12

# Audio processing
pydub==0.25.1
numpy==1.26.4

# Utilities
python-dotenv==1.0.1
